Revert "OpenGL: Single texture draw, no multitexture"
This reverts commit a2ce827b459a4e4806813363b0be9f2c9f0473d5.

+++ Furball.Vixie.Backends.Shared/Vertex.cs
@@ -6,5 +6,5 @@ public struct Vertex {
     public Vector2 Position;
     public Vector2 TextureCoordinate;
     public Color   Color;
-    public long    TexId;
+    public int     TexId;
 }
